+++ webgui.py
@@ -22,7 +22,7 @@ from modules.gui.gui_json_util import GuiJsonUtil
 from handlers.json_handler import JsonHandler
 
 from models.kaiseki import Kaiseki
-from models.kijiku import Kijiku
+from modules.common.translation import Kijiku
 
 from kudasai import Kudasai
